feat: add LogoutPop component and integrate logout functionality in Header
