+++ client/src/layout/Header.jsx
@@ -6,26 +6,30 @@ import {
     IconButton, Menu, Avatar
 } from '@mui/material';
 import MenuIcon from '@mui/icons-material/Menu';
-import HomeIcon from '@mui/icons-material/Home';
 import CloseIcon from '@mui/icons-material/Close';
-import PersonIcon from '@mui/icons-material/Person';
+import HomeIcon from '@mui/icons-material/Home';
+import LogoutIcon from '@mui/icons-material/Logout';
 import DescriptionIcon from '@mui/icons-material/Description';
+import AccountCircleIcon from '@mui/icons-material/AccountCircle';
 import EnhancedEncryptionIcon from '@mui/icons-material/EnhancedEncryption';
 import ConnectWithoutContactIcon from '@mui/icons-material/ConnectWithoutContact';
 
-import './Header.css';
-import LogoutPop from './LogoutPop';
-import AuthProvider from '../../middleware/AuthProvider';
 
+import './Header.css';
+import { useAuth } from '../hooks/useAuth';
+import LogoutPop from '../components/LogoutPop';
+
+
+// Header component
 export default function Header() {
     const navigate = useNavigate();
     const location = useLocation();
-    const { userName } = AuthProvider();
+    const { userData } = useAuth();
     const [open, setOpen] = useState(false);
     const [popUser, setPopUser] = useState(null);
     const [openLogout, setOpenLogout] = useState(false);
 
-    const isActive = (page) => location.pathname === '/' + page;
+    const isActive = (page) => location.pathname.split('/')[1] === page;
 
     const toggleDrawer = (page) => {
         setOpen(!open);
@@ -43,45 +47,40 @@ export default function Header() {
         if (setting === 'logout') setOpenLogout(true);
     };
 
-
     return (
         <AppBar position="fixed">
             {openLogout && <LogoutPop openLogout={openLogout} setOpenLogout={setOpenLogout} />}
             <Container maxWidth="lg">
-                <Toolbar disableGutters>
-                    <Tooltip arrow title="Secure Vault Application" placement="bottom">
-                        <Avatar variant="square" alt="header-icon" src="header-icon.png" sx={{ display: { xs: 'none', md: 'flex' } }} />
+                <Toolbar disableGutters variant="dense">
+                    <Tooltip arrow placement="bottom" title="Click to refresh">
+                        <Avatar
+                            src="/header-icon.png"
+                            variant="square"
+                            alt="header-icon"
+                            onClick={() => { window.location.reload() }}
+                            sx={{ display: { xs: 'none', md: 'flex', cursor: 'pointer' } }}
+                        />
                     </Tooltip>
-                    <Typography noWrap variant="h5" fontWeight={600} sx={{ display: { xs: 'none', md: 'flex' } }}>
+                    <Typography noWrap variant="h6" sx={{ display: { xs: 'none', md: 'flex' } }}>
                         &nbsp;Secure Vault
                     </Typography>
 
                     <Box sx={{ flexGrow: 1, display: { xs: 'none', md: 'flex' } }} ml={3}>
-                        <Tooltip arrow title="Home Page" placement="bottom">
-                            <MenuItem onClick={() => navigate('/home')} className={isActive('home') ? "active-route" : "non-active-route"}>
-                                <HomeIcon />&nbsp;<Typography variant="body1">Home</Typography>
-                            </MenuItem>
-                        </Tooltip>
-                        <Tooltip arrow title="Your Password Vault" placement="bottom">
-                            <MenuItem onClick={() => navigate('/vault')} className={isActive('vault') ? "active-route" : "non-active-route"}>
-                                <EnhancedEncryptionIcon />&nbsp;<Typography variant="body1">Vault</Typography>
-                            </MenuItem>
-                        </Tooltip>
-                        <Tooltip arrow title="Your Personal Journal" placement="bottom">
-                            <MenuItem onClick={() => navigate('/journal')} className={isActive('journal') ? "active-route" : "non-active-route"}>
-                                <DescriptionIcon />&nbsp;<Typography variant="body1">Journal</Typography>
-                            </MenuItem>
-                        </Tooltip>
-                        <Tooltip arrow title="Your Account Details" placement="bottom">
-                            <MenuItem onClick={() => navigate('/account')} className={isActive('account') ? "active-route" : "non-active-route"}>
-                                <PersonIcon />&nbsp;<Typography variant="body1">MyAccount</Typography>
-                            </MenuItem>
-                        </Tooltip>
-                        <Tooltip arrow title="Connect with Developer" placement="bottom">
-                            <MenuItem onClick={() => navigate('/collaborate')} className={isActive('collaborate') ? "active-route" : "non-active-route"}>
-                                <ConnectWithoutContactIcon />&nbsp;<Typography variant="body1">Collaborate</Typography>
-                            </MenuItem>
-                        </Tooltip>
+                        <MenuItem onClick={() => navigate('/home')} className={isActive('home') ? "active-route" : "non-active-route"}>
+                            <HomeIcon />&nbsp;<Typography variant="body1">Home</Typography>
+                        </MenuItem>
+                        <MenuItem onClick={() => navigate('/vault')} className={isActive('vault') ? "active-route" : "non-active-route"}>
+                            <EnhancedEncryptionIcon />&nbsp;<Typography variant="body1">Vault</Typography>
+                        </MenuItem>
+                        <MenuItem onClick={() => navigate('/journal')} className={isActive('journal') ? "active-route" : "non-active-route"}>
+                            <DescriptionIcon />&nbsp;<Typography variant="body1">Journal</Typography>
+                        </MenuItem>
+                        <MenuItem onClick={() => navigate('/account')} className={isActive('account') ? "active-route" : "non-active-route"}>
+                            <AccountCircleIcon />&nbsp;<Typography variant="body1">Account</Typography>
+                        </MenuItem>
+                        <MenuItem onClick={() => navigate('/collaborate')} className={isActive('collaborate') ? "active-route" : "non-active-route"}>
+                            <ConnectWithoutContactIcon />&nbsp;<Typography variant="body1">Collaborate</Typography>
+                        </MenuItem>
                     </Box>
 
                     <Box sx={{ flexGrow: 1, display: { xs: 'flex', md: 'none' } }}>
@@ -94,15 +93,15 @@ export default function Header() {
                         </Button>
                     </Box>
 
-                    <Avatar variant="square" alt="header-icon" src="header-icon.png" sx={{ display: { xs: 'flex', md: 'none' } }} />
+                    <Avatar variant="square" alt="header-icon" src="/header-icon.png" sx={{ display: { xs: 'flex', md: 'none' } }} />
                     <Typography noWrap variant="h5" sx={{ display: { xs: 'flex', md: 'none' }, flexGrow: 1 }}>
                         &nbsp;Secure Vault
                     </Typography>
 
                     <Box sx={{ flexGrow: 0 }}>
-                        <Tooltip arrow title="Logout User" placement="bottom">
+                        <Tooltip arrow placement="bottom" title="Click to logout">
                             <IconButton onClick={handleOpenUserMenu} sx={{ p: 0 }}>
-                                <Avatar alt={userName} src={`https://robohash.org/${userName}`} className='profileAvt' />
+                                <Avatar alt={userData?.name} src={`https://robohash.org/${userData?.name}`} className='profileAvt' />
                             </IconButton>
                         </Tooltip>
                         <Menu
@@ -122,7 +121,9 @@ export default function Header() {
                             onClose={handleCloseUserMenu}
                         >
                             <MenuItem onClick={() => handleCloseUserMenu("logout")}>
-                                <Typography sx={{ textAlign: 'center' }}>Logout</Typography>
+                                <Typography sx={{ display: 'flex', textAlign: 'center', fontWeight: 600 }}>
+                                    Logout &nbsp; <LogoutIcon color='secondary' />
+                                </Typography>
                             </MenuItem>
                         </Menu>
                     </Box>
@@ -150,7 +151,7 @@ export default function Header() {
                             <DescriptionIcon />&nbsp;<Typography variant="body1" fontWeight={800}>Journal</Typography>
                         </MenuItem>
                         <MenuItem onClick={() => toggleDrawer('account')} className={isActive('account') ? "pop-active" : "pop-non-active"}>
-                            <PersonIcon />&nbsp;<Typography variant="body1" fontWeight={800}>MyAccount</Typography>
+                            <AccountCircleIcon />&nbsp;<Typography variant="body1" fontWeight={800}>Account</Typography>
                         </MenuItem>
                         <MenuItem onClick={() => toggleDrawer('collaborate')} className={isActive('collaborate') ? "pop-active" : "pop-non-active"}>
                             <ConnectWithoutContactIcon />&nbsp;<Typography variant="body1" fontWeight={800}>Collaborate</Typography>
